Display an explicit error if the generation of a package failed

+++ src/60_system_gentoo/10_packages_gentoo-variant.sh
@@ -152,19 +152,31 @@ gentoo_package_build_single() {
 	esac
 
 	# Run the actual package generation, using ebuild
+	local package_generation_return_code
 	information_package_building "$package_name"
 	mkdir --parents "${PLAYIT_WORKDIR}/portage-tmpdir"
 	local ebuild_path
 	ebuild_path=$(gentoo_ebuild_path "$package")
 	ebuild "$ebuild_path" manifest 1>/dev/null
-
 	if [ -n "$binkpg_compress" ]; then
 		debug_external_command "PORTAGE_TMPDIR=\"${PLAYIT_WORKDIR}/portage-tmpdir\" PKGDIR=\"${PLAYIT_WORKDIR}/gentoo-pkgdir\" BINPKG_COMPRESS=\"$binkpg_compress\" fakeroot -- ebuild \"$ebuild_path\" package 1>/dev/null"
+		set +o errexit
 		PORTAGE_TMPDIR="${PLAYIT_WORKDIR}/portage-tmpdir" PKGDIR="${PLAYIT_WORKDIR}/gentoo-pkgdir" BINPKG_COMPRESS="$binkpg_compress" fakeroot -- ebuild "$ebuild_path" package 1>/dev/null
+		package_generation_return_code=$?
+		set -o errexit
 	else
 		debug_external_command "PORTAGE_TMPDIR=\"${PLAYIT_WORKDIR}/portage-tmpdir\" PKGDIR=\"${PLAYIT_WORKDIR}/gentoo-pkgdir\" fakeroot -- ebuild \"$ebuild_path\" package 1>/dev/null"
+		set +o errexit
 		PORTAGE_TMPDIR="${PLAYIT_WORKDIR}/portage-tmpdir" PKGDIR="${PLAYIT_WORKDIR}/gentoo-pkgdir" fakeroot -- ebuild "$ebuild_path" package 1>/dev/null
+		package_generation_return_code=$?
+		set -o errexit
 	fi
+
+	if [ $package_generation_return_code -ne 0 ]; then
+		error_package_generation_failed "$package_name"
+		return 1
+	fi
+
 	mkdir --parents "$generated_package_directory"
 	mv "${PLAYIT_WORKDIR}/gentoo-pkgdir/games-playit/${package_name}" "$generated_package_path"
 	rm --recursive "${PLAYIT_WORKDIR}/portage-tmpdir"
